+++ Project/Assets/Prefab/Character/Player/Player_Character_(Beta).prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f31b5c0675d9596439d3d94af787a680
+guid: 27f3d4383cdea134f862cccec3e8195e
 PrefabImporter:
   externalObjects: {}
   userData:
