+++ Project/Assets/Script/Object/Animation/AnimationMovementController.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 58c7d903b4b7496a82991b48633ba81b
+timeCreated: 1761127150
