+++ Project/Assets/Script/Object/Character/Player/Animation/RollingStateMachine.cs
@@ -4,8 +4,6 @@ namespace Backend.Object.Character.Player.Animation
 {
     public class RollingStateMachine : StateMachineBehaviour
     {
-        public AnimationCurve speed;
-
         private AdvancedActionController _controller;
 
         public override void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
@@ -15,16 +13,8 @@ namespace Backend.Object.Character.Player.Animation
             _controller?.OnRollingStateEntered();
         }
 
-        public override void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
-        {
-            var normalized = stateInfo.normalizedTime % 1f;
-
-            _controller.deltaSpeed = speed.Evaluate(normalized);
-        }
-
         public override void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
         {
-            _controller.deltaSpeed = 1f;
             _controller.OnRollingStateExited();
         }
     }
